Migrate DevLights Project page to new modularized structure
