--- src/components/DevLights/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./DevLights";
