+++ src/modules/core/components/Router.tsx
@@ -1,11 +1,11 @@
 import { BrowserRouter, Redirect, Route, Switch } from "react-router-dom";
-import DevLights from "../../../components/DevLights";
 import Footer from "../components/Footer";
 import Projects from "../../projects/components/Projects";
 import TECKboard from "../../../components/TECKboard";
 import EmptyPage from "../design-system/EmptyPage";
 import Menu from "./Menu";
 import Home from "../../home/components/Home";
+import DevLights from "../../devlights/components/DevLights";
 
 export default function Router() {
   return (
@@ -20,10 +20,10 @@ export default function Router() {
         <Route path="/projects/home">
           <Projects />
         </Route>
-        <Route exact path="/project/teckboard">
+        <Route exact path="/projects/teckboard">
           <TECKboard />
         </Route>
-        <Route exact path="/project/devlights">
+        <Route exact path="/projects/devlights">
           <DevLights />
         </Route>
