+++ src/modules/core/design-system/TimelineItem.tsx
@@ -1,12 +1,14 @@
 import {
+  TimelineConnector,
   TimelineContent,
+  TimelineDot,
   TimelineItem,
   TimelineOppositeContent,
   TimelineSeparator,
 } from "@mui/lab";
-import Tilt from "react-parallax-tilt";
 import { styled, Typography } from "@mui/material";
 import React from "react";
+import Tilt from "react-parallax-tilt";
 import { contrastBlack } from "../../../modules/core/utils/textContrast";
 
 const PREFIX = "CustomTimelineItem";
@@ -19,9 +21,10 @@ const classes = {
   textItem: `${PREFIX}-textItem`,
 };
 
-const StyledCustomTimelineItem = styled("div")(({ theme }) => ({
+const StyledCustomTimelineItem = styled(TimelineItem)(({ theme }) => ({
   [`&.${classes.root}`]: {
     flex: 3,
+    marginTop: theme.spacing(8),
   },
   [`& .${classes.item}`]: {
     margin: theme.spacing(8),
@@ -32,56 +35,67 @@ const StyledCustomTimelineItem = styled("div")(({ theme }) => ({
     fontFamily: "bungee",
     textAlign: "center",
     alignSelf: "center",
+    width: "50%",
     color: contrastBlack(theme.palette.background.default) ? "#000" : "#fff",
   },
   [`& .${classes.text}`]: {
     fontSize: "2.0vh",
+    width: "50%",
     textAlign: "center",
     color: contrastBlack(theme.palette.background.default) ? "#000" : "#fff",
   },
   [`& .${classes.textItem}`]: {
     flex: 5,
+    display: "flex",
+    justifyContent: "center",
     justifyItems: "center",
+    textAlign: "center",
   },
 }));
 
 export interface TimelineItemProps {
   key: string;
-  title: string;
+  title?: string;
   description: string | React.ReactNode;
-  image: string;
+  image?: string;
   imageStyle?: React.CSSProperties;
   rotateImage?: boolean;
   direction: "normal" | "opposite";
   textChildren?: React.ReactNode;
+  useDot?: boolean;
 }
 
 export default function CustomizedTimelineItem(props: TimelineItemProps) {
-  if (props.direction === "normal") {
-    return <CustomTimelineItem {...props} />;
-  } else {
-    return <OppositeCustomTimelineItem {...props} />;
-  }
+  return <CustomTimelineItem {...props} />;
 }
 
 function CustomTimelineItem(props: TimelineItemProps) {
   return (
     <StyledCustomTimelineItem className={classes.item}>
-      <TimelineItem className={classes.item} key={props.key}>
-        <TimelineOppositeContent className={classes.textItem}>
-          <div
-            style={{ width: "50%", display: "flex", flexDirection: "column" }}
-          >
-            <h1 className={classes.title}>{props.title}</h1>
+      <TimelineOppositeContent className={classes.textItem}>
+        <div
+          style={{
+            width: "100%",
+            display: "flex",
+            flexDirection: "column",
+            alignItems: "center",
+          }}
+        >
+          <h1 className={classes.title}>{props.title}</h1>
 
-            <Typography className={classes.text}>
-              {props.description}
-            </Typography>
-            {props.textChildren}
-          </div>
-        </TimelineOppositeContent>
-        <TimelineSeparator sx={{ flex: 1 }} />
-        <TimelineContent sx={{ flex: 5 }}>
+          <Typography className={classes.text}>{props.description}</Typography>
+          {props.textChildren}
+        </div>
+      </TimelineOppositeContent>
+      {props.useDot && (
+        <TimelineSeparator>
+          <TimelineDot />
+          <TimelineConnector />
+        </TimelineSeparator>
+      )}
+
+      <TimelineContent sx={{ flex: 5 }}>
+        {props.image != null ? (
           <Tilt
             style={{
               width: "auto",
@@ -103,50 +117,10 @@ function CustomTimelineItem(props: TimelineItemProps) {
               alt={props.title}
             />
           </Tilt>
-        </TimelineContent>
-      </TimelineItem>
-    </StyledCustomTimelineItem>
-  );
-}
-
-function OppositeCustomTimelineItem(props: TimelineItemProps) {
-  return (
-    <StyledCustomTimelineItem className={classes.item}>
-      <TimelineItem key={props.key} className={classes.item}>
-        <TimelineOppositeContent sx={{ flex: 5 }}>
-          <Tilt
-            style={{
-              width: "auto",
-              height: "auto",
-            }}
-          >
-            <img
-              style={{
-                borderRadius: 10,
-                transform: props.rotateImage
-                  ? "rotate(-0.025turn) translate(20%, 10%)"
-                  : undefined,
-              }}
-              height={props.imageStyle?.height ?? "65%"}
-              width={props.imageStyle?.width ?? "65%"}
-              src={props.image}
-              alt={props.title}
-            />
-          </Tilt>
-        </TimelineOppositeContent>
-        <TimelineSeparator sx={{ flex: 1 }} />
-
-        <TimelineContent className={classes.textItem}>
-          <div style={{ width: "50%" }}>
-            <h1 className={classes.title}>{props.title}</h1>
-            <Typography className={classes.text}>
-              {props.description}
-            </Typography>
-            {props.textChildren}
-          </div>
-        </TimelineContent>
-        <TimelineSeparator />
-      </TimelineItem>
+        ) : (
+          <></>
+        )}
+      </TimelineContent>
     </StyledCustomTimelineItem>
   );
 }
